+++ systemd/start.sh
@@ -1,2 +1,3 @@
 #!/bin/sh
-PORT=8505 ADMIN=hkimura PG_USER=postgres PG_PASS=password java -jar konpy.jar >> log/konpy.log 2>> log/error.log
+# PORT=8505 ADMIN=hkimura PG_USER=postgres PG_PASS=password java -jar konpy.jar >> log/konpy.log 2>> log/error.log
+java -jar konpy.jar >> log/konpy.log 2>> log/error.log